deleted little bug
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@255 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/pss_full/pss_pss.c
@@ -867,7 +867,7 @@ if (err != sizename) dserror("error reading pss-file");
 if ( strncmp(name,test_name,strlen(name)) != 0 )
 {
 /*-------------------- not the right record, so jump it and start again */
-   err = fread(&handle_dummy,sizeof(int),1,allfiles.out_pss);
+   err = fread(&handle_dummy,sizeof(long int),1,allfiles.out_pss);
    if (err != 1) dserror("error reading pss-file");
    err=fread(dimensions,sizeof(int),3,allfiles.out_pss);
    if (err != 3) dserror("error reading pss-file");
